Remove constraint node interface
