+++ generator/src/main/java/com/scottlogic/deg/generator/decisiontree/DecisionTreeSimplifier.java
@@ -77,7 +77,7 @@ public class DecisionTreeSimplifier {
                     }
                 },
                 (node1, node2) ->
-                    new TreeConstraintNode(
+                    new ConstraintNode(
                         Stream
                             .concat(node1.getAtomicConstraints().stream(), node2.getAtomicConstraints().stream())
                             .collect(Collectors.toList()),
