+++ generator/src/main/java/com/scottlogic/deg/generator/decisiontree/MostProlificConstraintOptimiser.java
@@ -67,8 +67,8 @@ public class MostProlificConstraintOptimiser implements DecisionTreeOptimiser {
         }
 
         // Add most prolific constraint to new decision node
-        ConstraintNode factorisingConstraintNode = new TreeConstraintNode(mostProlificAtomicConstraint);
-        ConstraintNode negatedFactorisingConstraintNode = new TreeConstraintNode(negatedMostProlificConstraint);
+        ConstraintNode factorisingConstraintNode = new ConstraintNode(mostProlificAtomicConstraint);
+        ConstraintNode negatedFactorisingConstraintNode = new ConstraintNode(negatedMostProlificConstraint);
 
         Set<ConstraintNode> otherOptions = new HashSet<>();
         Set<DecisionNode> decisionsToRemove = new HashSet<>();
