+++ generator/src/main/java/com/scottlogic/deg/generator/walker/reductive/PrunedConstraintState.java
@@ -20,7 +20,6 @@ import com.scottlogic.deg.common.profile.Field;
 import com.scottlogic.deg.common.profile.constraints.atomic.AtomicConstraint;
 import com.scottlogic.deg.generator.decisiontree.ConstraintNode;
 import com.scottlogic.deg.generator.decisiontree.DecisionNode;
-import com.scottlogic.deg.generator.decisiontree.TreeConstraintNode;
 import com.scottlogic.deg.generator.fieldspecs.FieldSpec;
 
 import java.util.ArrayList;
@@ -55,7 +54,7 @@ class PrunedConstraintState {
     }
 
     ConstraintNode getNewConstraintNode() {
-        return new TreeConstraintNode(newAtomicConstraints, newDecisionNodes);
+        return new ConstraintNode(newAtomicConstraints, newDecisionNodes);
     }
 
     Map<Field, FieldSpec> addPulledUpFieldsToMap(Map<Field, FieldSpec> previousFieldSpecs) {
