+++ generator/src/test/java/com/scottlogic/deg/generator/decisiontree/MaxStringLengthInjectingDecisionTreeFactoryTests.java
@@ -43,7 +43,7 @@ class MaxStringLengthInjectingDecisionTreeFactoryTests {
         Profile profile = mock(Profile.class);
         ProfileFields fields = new ProfileFields(Collections.singletonList(new Field("field 1")));
         DecisionTree underlyingTree = new DecisionTree(
-            new TreeConstraintNode(Collections.emptySet(), Collections.emptySet()),
+            new ConstraintNode(Collections.emptySet(), Collections.emptySet()),
             fields
         );
         when(underlyingFactory.analyse(profile)).thenReturn(underlyingTree);
