+++ generator/src/test/java/com/scottlogic/deg/generator/decisiontree/treepartitioning/ConstraintToFieldMapperTests.java
@@ -22,7 +22,7 @@ import com.scottlogic.deg.common.profile.constraints.atomic.IsInSetConstraint;
 import com.scottlogic.deg.common.profile.constraints.atomic.AtomicConstraint;
 import com.scottlogic.deg.generator.decisiontree.DecisionNode;
 import com.scottlogic.deg.generator.decisiontree.DecisionTree;
-import com.scottlogic.deg.generator.decisiontree.TreeConstraintNode;
+import com.scottlogic.deg.generator.decisiontree.ConstraintNode;
 import com.scottlogic.deg.generator.decisiontree.TreeDecisionNode;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.DistributedSet;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.WeightedElement;
@@ -58,7 +58,7 @@ class ConstraintToFieldMapperTests {
 
         final AtomicConstraint constraint = new IsInSetConstraint(new Field("B"), whitelistOf("test-value"));
         final DecisionNode decision = new TreeDecisionNode(
-            new TreeConstraintNode(constraint));
+            new ConstraintNode(constraint));
 
         givenDecisions(decision);
 
@@ -90,24 +90,24 @@ class ConstraintToFieldMapperTests {
         final AtomicConstraint constraintF = new IsInSetConstraint(new Field("F"), whitelistOf("test-value"));
 
         final DecisionNode decisionABC = new TreeDecisionNode(
-            new TreeConstraintNode(
+            new ConstraintNode(
                 Collections.emptyList(),
                 Arrays.asList(
-                    new TreeDecisionNode(new TreeConstraintNode(constraintA)),
-                    new TreeDecisionNode(new TreeConstraintNode(constraintB)),
-                    new TreeDecisionNode(new TreeConstraintNode(constraintC))
+                    new TreeDecisionNode(new ConstraintNode(constraintA)),
+                    new TreeDecisionNode(new ConstraintNode(constraintB)),
+                    new TreeDecisionNode(new ConstraintNode(constraintC))
                 )
             )
         );
 
         final DecisionNode decisionDEF = new TreeDecisionNode(
-            new TreeConstraintNode(
+            new ConstraintNode(
                 Collections.emptyList(),
                 Collections.singletonList(
                     new TreeDecisionNode(
-                        new TreeConstraintNode(constraintD),
-                        new TreeConstraintNode(constraintE),
-                        new TreeConstraintNode(constraintF))
+                        new ConstraintNode(constraintD),
+                        new ConstraintNode(constraintE),
+                        new ConstraintNode(constraintF))
                 )
             )
         );
@@ -149,7 +149,7 @@ class ConstraintToFieldMapperTests {
     private void getMappings() {
         mappings = new ConstraintToFieldMapper()
             .mapConstraintsToFields(new DecisionTree(
-                new TreeConstraintNode(constraintsList, decisionsList),
+                new ConstraintNode(constraintsList, decisionsList),
                 fields
             ));
     }
