+++ generator/src/test/java/com/scottlogic/deg/generator/decisiontree/treepartitioning/RelatedFieldTreePartitionerTests.java
@@ -34,8 +34,8 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 class RelatedFieldTreePartitionerTests {
-    private static final TreeConstraintNode emptyConstraint
-        = new TreeConstraintNode(Collections.emptySet(), Collections.emptySet());
+    private static final ConstraintNode emptyConstraint
+        = new ConstraintNode(Collections.emptySet(), Collections.emptySet());
 
     @Test
     void shouldSplitTreeIntoPartitions() {
@@ -259,7 +259,7 @@ class RelatedFieldTreePartitionerTests {
     }
 
     private ConstraintNode constraint(String[] fieldNames, DecisionNode... decisions) {
-        return new TreeConstraintNode(
+        return new ConstraintNode(
             Stream.of(fieldNames)
                 .map(this::atomicConstraint)
                 .collect(Collectors.toList()),
